RF/ENH: Move `StaticPeriod` to clock.py
We still import it in `psychopy.core` for backward-compat.
